Add support for non-deflate compression.
While we cannot reasonably support the compression and decompression of entries using all formats that zip supports we can pass through entries that are already compressed. This change converts the compression flag to an integer value representative of the compressio type, as well as adding an assertion when attempting to get the blob for an entry if we cannot decode it.