Break up archive parsing to prevent blocks
The required time for creating the blobs for the entries from the source blob can be quite long, likely due to some underlying memory copy being required. Hence to prevent a long lock up this will suspend the task for a short duration if a certain duration has passed since the current task has started, allowing for other tasks to execute .

Also resolves a small issue reading the field of a entry which could cause issues.