bugfix - some filenames are incorrectly treated as invalid
Changes the path verification function to use a blacklist of 2 characters instead of a whitelist. The 2 invalid characters are "/" and NUL. In practise the slash is ignored because we split the path by slashes just prior to this. But it's good to include in case this function is ever rearranged somehow.