Change `.files` iterator to return a static list
It's fairly trivial to implement a system to copy, modify or rename various files within the archive using the files iterator. Unfortunately the dynamic nature of the Map.entries iterator can lead to an infinite loop using this pattern because additions are allways visited by the iterator.

This is a simple change to take a snapshot of the entries to iterate over, enabling a simpler use for that pattern.