fix: potential issue with task splitting in `from_blob`
The method for parsing a zip file can take a long time to run, hence to prevent it blocking the thread for an extended time if it exceeds the configured task duration ( 32ms at the moment ) it will pause for a short time.

To decide how long the task has been running we had been using Date.now(). However, this is based on system time. Hence under some situations it can actually decrease. Additionally in some environments ( like cloudflare workers ) it will always return the same value.

Hence this change introduces 2 new ways of determining the duration of the task. The preferential one is to use `performance.now` which is a non-decreasing high resolution time stamp. In the situation that it isn't present we use Date.now, but we check the return value against the previous value and if it is less than or equal to the previous we just increment the previous value by 0.1 instead. This has the side effect that each fragment of a task is presumed to take at least 0.1ms if performance.now isn't present. It's possible that this will increase the number of pauses we experience, but still allows for reading 320 entries per task.