add: `move` and `get_folder` methods
This change adds 2 new methods to the ZipArchive class.

The first is quite a simple method `get_folder` as the name suggests this gets a folder entry from the archive. While the standard `get` method can still get folder entries this normalises the file path to a directory path.

The second is conceptually simple, but due to path manipulation quite complicated to implement. ZipEntry objects do not contain their file name, this is a purposeful decision as it allows for the same entry to be used for more than one file if the contents is identical. Hence we can rename an entry by just moving it to another location in the archive. This is a lot faster than reading and setting a new object as it skips the decompress, compress steps.