add: clone method for ZipEntry objects
Method has been added for the "copy" method on the archive. Cloning creates a new ZipEntry with identical data to the old one. Mutable attributes are copied as well to ensure that the ZipEntry is not sharing the values, so they act uniquely. However, immutable data ( such as the backing object ) is not copied which allows for some optimisation.