add: `copy` method on ZipArchive
Utilising the ZipEntry `clone` method this method behaves similarly to the recently added `move` method with 2 key differences. Firstly the source entry is not deleted. Secondly the new entry is not the same as the source, so it's possible to modify the new entry without affecting the old one. The entries still share an immutable backing object which allows for some optimisation. Ideally we would have some sort of COW design here to reduce the overhead of the copy even further.